docs: added `KeyboardAvoidingView` and `KeyboardStickyView` comparison
